Custom Block tests

+++ BlocksToCpp/testsForOutputCCode/equal_cCode.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 using namespace std;
 int main() {
-   double item = 54;
+   int item = 54;
   if(item == 54) {
     cout << ("EQUAL") << endl;
   };
