+++ BlocksToCpp/testsForOutputCCode/loopToPrint40_cCode.c
@@ -3,8 +3,8 @@
 #include <stdlib.h>
 using namespace std;
 int main() {
-   double item = 0;
-  while(item % 10 != 0) {
+   int item = 0;
+  while((int)item % (int)10 != 0) {
     int item = item + 4;
   };
   cout << (item) << endl;
