+++ BlocksToCpp/testsForOutputCCode/printSquared_cCode.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 using namespace std;
 int main() {
-   double item = 1;
+   int item = 1;
   for(int i = 1; i<=(10); i+=(1)) {
     cout << (      pow(i, 2)) << endl;
    };
