+++ BlocksToCpp/testsForOutputCCode/printTill25_cCode.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 using namespace std;
 int main() {
-   double item = 1;
+   int item = 1;
   for(int i = 1; i<=(100); i+=(1)) {
     cout << (i) << endl;
    };
